services/api/app/requirements.txt
sanic
sanic-cors>=0.10.0.b1
sanic-validation
sanic-openapi
arq
authlib
sentry_sdk
tortoise-orm
asyncpg
aiohttp
apscheduler
hashlib